TYPE ALIASES: WIN32 STYLE

+++ lib/tklib/src/tklib/formatter.cpp
@@ -32,19 +32,18 @@ namespace tklib
      *  DATE    : 2026                                                               *
      *  PURPOSE : Format val as a zero-padded 16-digit hex string into buf.          *
      ********************************************************************************/
-    void fmt_hex(char *buf, usize size, u64 val) noexcept
+    void fmt_hex(char *buf, SIZE_T size, QWORD val) noexcept
     {
         if (!buf || size < 19)
             return;
 
-        buf[0] = '0';
-        buf[1] = 'x';
+        buf[0]  = '0';
+        buf[1]  = 'x';
         buf[18] = '\0';
 
-        for (i32 i = 17; i >= 2; --i)
-        {
-            buf[i] = HEX_CHARS[val & 0xF];
-            val >>= 4;
+        for (LONG i = 17; i >= 2; --i) {
+            buf[i]   = HEX_CHARS[val & 0xF];
+            val    >>= 4;
         }
     }
 
@@ -54,19 +53,18 @@ namespace tklib
      *  DATE    : 2026                                                               *
      *  PURPOSE : Format val as a zero-padded 8-digit hex string into buf.           *
      ********************************************************************************/
-    void fmt_hex32(char *buf, usize size, u32 val) noexcept
+    void fmt_hex32(char *buf, SIZE_T size, DWORD val) noexcept
     {
         if (!buf || size < 11)
             return;
 
-        buf[0] = '0';
-        buf[1] = 'x';
+        buf[0]  = '0';
+        buf[1]  = 'x';
         buf[10] = '\0';
 
-        for (i32 i = 9; i >= 2; --i)
-        {
-            buf[i] = HEX_CHARS[val & 0xF];
-            val >>= 4;
+        for (LONG i = 9; i >= 2; --i) {
+            buf[i]   = HEX_CHARS[val & 0xF];
+            val    >>= 4;
         }
     }
 
@@ -76,29 +74,27 @@ namespace tklib
      *  DATE    : 2026                                                               *
      *  PURPOSE : Format val as a decimal string into buf.                           *
      ********************************************************************************/
-    void fmt_dec(char *buf, usize size, u64 val) noexcept
+    void fmt_dec(char *buf, SIZE_T size, QWORD val) noexcept
     {
         if (!buf || size < 2)
             return;
 
-        if (val == 0)
-        {
+        if (val == 0) {
             buf[0] = '0';
             buf[1] = '\0';
             return;
         }
 
         char tmp[21];
-        i32 i = 20;
+        LONG i  = 20;
         tmp[20] = '\0';
 
-        while (val > 0 && i > 0)
-        {
-            tmp[--i] = '0' + static_cast<char>(val % 10);
-            val /= 10;
+        while (val > 0 && i > 0) {
+            tmp[--i]  = '0' + static_cast<char>(val % 10);
+            val      /= 10;
         }
 
-        usize len = static_cast<usize>(20 - i);
+        SIZE_T len = static_cast<SIZE_T>(20 - i);
         if (len >= size)
             len = size - 1;
 
@@ -112,19 +108,18 @@ namespace tklib
      *  DATE    : 2026                                                               *
      *  PURPOSE : Format val as a signed decimal string into buf.                    *
      ********************************************************************************/
-    void fmt_dec_signed(char *buf, usize size, i64 val) noexcept
+    void fmt_dec_signed(char *buf, SIZE_T size, LONGLONG val) noexcept
     {
         if (!buf || size < 2)
             return;
 
-        if (val >= 0)
-        {
-            fmt_dec(buf, size, static_cast<u64>(val));
+        if (val >= 0) {
+            fmt_dec(buf, size, static_cast<QWORD>(val));
             return;
         }
 
         buf[0] = '-';
-        fmt_dec(buf + 1, size - 1, static_cast<u64>(-val));
+        fmt_dec(buf + 1, size - 1, static_cast<QWORD>(-val));
     }
 
     /* *******************************************************************************
@@ -133,19 +128,18 @@ namespace tklib
      *  DATE    : 2026                                                               *
      *  PURPOSE : Format val as a zero-padded 64-bit binary string into buf.         *
      ********************************************************************************/
-    void fmt_bin(char *buf, usize size, u64 val) noexcept
+    void fmt_bin(char *buf, SIZE_T size, QWORD val) noexcept
     {
         if (!buf || size < 67)
             return;
 
-        buf[0] = '0';
-        buf[1] = 'b';
+        buf[0]  = '0';
+        buf[1]  = 'b';
         buf[66] = '\0';
 
-        for (i32 i = 65; i >= 2; --i)
-        {
-            buf[i] = '0' + static_cast<char>(val & 1);
-            val >>= 1;
+        for (LONG i = 65; i >= 2; --i) {
+            buf[i]   = '0' + static_cast<char>(val & 1);
+            val    >>= 1;
         }
     }
 
@@ -155,33 +149,26 @@ namespace tklib
      *  DATE    : 2026                                                               *
      *  PURPOSE : Format val as a human-readable size string (KB, MB, GB).           *
      ********************************************************************************/
-    void fmt_size(char *buf, usize size, u64 val) noexcept
+    void fmt_size(char *buf, SIZE_T size, QWORD val) noexcept
     {
         if (!buf || size < 2)
             return;
 
         const char *unit;
-        u64 display;
+        QWORD display;
 
-        if (val >= 1024ull * 1024ull * 1024ull)
-        {
+        if (val >= 1024ull * 1024ull * 1024ull) {
             display = val / (1024ull * 1024ull * 1024ull);
-            unit = " GB";
-        }
-        else if (val >= 1024ull * 1024ull)
-        {
+            unit    = " GB";
+        } else if (val >= 1024ull * 1024ull) {
             display = val / (1024ull * 1024ull);
-            unit = " MB";
-        }
-        else if (val >= 1024ull)
-        {
+            unit    = " MB";
+        } else if (val >= 1024ull) {
             display = val / 1024ull;
-            unit = " KB";
-        }
-        else
-        {
+            unit    = " KB";
+        } else {
             display = val;
-            unit = " B";
+            unit    = " B";
         }
 
         char tmp[21];
